convert chaptermark timestamps relative to start
Since the timestamps that we are get for the chaptermarks, are UNIX Timestamps.
We have to convert them in a time object suitable from H:M:S output of the time
relative to the start of the audio.

* pentapub.utils.time
Added seconds_to_time(seconds).
I've found no nicer way to convert a second in a time object which than can be
strftime('%H:%M:%S') until that we've this helper function

* pentapub.utils.chaptermarks
Added is_relative_time(timestamp).
Checks if the given timestamp is less then 5400, which are 90 minutes in seconds.
We should add a way to get the actual length of the audio production.

Added convert_timestamp(timestamp, relative_to, offset).
Converts the given timestamp in seconds in a time object using the
seconds_to_time helper function.
If the given timestamp is greater then the audio length, the timestamp will be
set reltive to the timestamp in the `relative_to` parameter.

audacity_markter_to_xml(marker, root, start_time, offset) need more arguments
passed to convert_timestamp.

audacity_to_podlove_chapters(filename, start_time, offset), if start_time is `0`
which is default, the timestamp of the first chaptermark entry will be used as
`relative_to` argument for `convert_timestamp`